+++ src/AppRouter.tsx
@@ -1,0 +1,125 @@
+import { Center } from '@chakra-ui/react'
+import { AnimatePresence, motion } from 'framer-motion'
+import { useEffect, useMemo } from 'react'
+import { FormProvider, useForm, useWatch } from 'react-hook-form'
+import { Route, Routes, useLocation, useSearchParams } from 'react-router'
+import { ChatwootButton } from 'components/Chatwoot'
+import { SelectPair } from 'components/SelectPair'
+import { Status } from 'components/Status/Status'
+import { TradeInput } from 'components/TradeInput'
+import type { SwapFormData } from 'types/form'
+
+const selectPair = <SelectPair />
+const tradeInput = <TradeInput />
+const status = <Status />
+
+const slideVariants = {
+  initial: { opacity: 0, x: 20 },
+  animate: { opacity: 1, x: 0 },
+  exit: { opacity: 0, x: -20 },
+}
+
+const fadeUpVariants = {
+  initial: { opacity: 0, y: 20 },
+  animate: { opacity: 1, y: 0 },
+  exit: { opacity: 0, y: -20 },
+}
+
+const motionWrapperSx = {
+  width: '100%',
+  maxWidth: '450px',
+  display: 'flex',
+  justifyContent: 'center',
+}
+
+const transition = {
+  type: 'tween',
+  ease: [0.25, 0.1, 0.25, 1],
+  duration: 0.3,
+}
+
+const getVariants = (pathname: string) => {
+  // Not sure what this does but this looks good on status screen so
+  if (pathname === '/status') return fadeUpVariants
+  return slideVariants
+}
+
+export const AppRouter = () => {
+  const [searchParams, setSearchParams] = useSearchParams()
+  const location = useLocation()
+
+  const defaultValues = useMemo(
+    () => ({
+      sellAsset: searchParams.get('sellAsset') || undefined,
+      buyAsset: searchParams.get('buyAsset') || undefined,
+      sellAmountCryptoBaseUnit: searchParams.get('sellAmountCryptoBaseUnit') || undefined,
+      destinationAddress: searchParams.get('destinationAddress') || undefined,
+      refundAddress: searchParams.get('refundAddress') || undefined,
+    }),
+    [searchParams],
+  )
+
+  const methods = useForm<SwapFormData>({
+    mode: 'all',
+    reValidateMode: 'onChange',
+    defaultValues,
+  })
+
+  const formValues = useWatch({ control: methods.control })
+
+  // Synchronize queryparams with form
+  useEffect(() => {
+    const params = new URLSearchParams(searchParams)
+
+    Object.entries(formValues || {}).forEach(([key, value]) => {
+      if (value) {
+        params.set(key, String(value))
+      } else {
+        params.delete(key)
+      }
+    })
+
+    setSearchParams(params, { replace: true })
+  }, [formValues, setSearchParams, searchParams])
+
+  // Synchronize form with queryparams too
+  useEffect(() => {
+    const currentParams = Object.fromEntries(searchParams.entries())
+    const hasParams = Object.keys(currentParams).length > 0
+
+    if (hasParams) {
+      const formData: Partial<SwapFormData> = {}
+      searchParams.forEach((value, key) => {
+        if (key in defaultValues) {
+          formData[key as keyof SwapFormData] = value
+        }
+      })
+      methods.reset(formData)
+    }
+  }, [searchParams, methods, defaultValues])
+
+  return (
+    <FormProvider {...methods}>
+      <Center width='full' height='100vh' flexDir='column'>
+        <AnimatePresence mode='wait' initial={false}>
+          <motion.div
+            key={location.pathname}
+            initial='initial'
+            animate='animate'
+            exit='exit'
+            variants={getVariants(location.pathname)}
+            transition={transition}
+            style={motionWrapperSx}
+          >
+            <Routes location={location}>
+              <Route path='/' element={selectPair} />
+              <Route path='/input' element={tradeInput} />
+              <Route path='/status' element={status} />
+            </Routes>
+          </motion.div>
+        </AnimatePresence>
+        <ChatwootButton />
+      </Center>
+    </FormProvider>
+  )
+}
